Testing: Enable no-unused-disable ESLint Comments plugin rule (#72940)
* Testing: Enable no-unused-disable ESLint Comments plugin rule

* Fix existing issues with no-unused-disable

* Remove "disable reason" comments that don't refer to anything

* Fix more formatting irregularities

* Ignore generated icon files for ESLint

These files are auto-generated, and expected to have some formatting issues. However, blanket disabling Prettier is incompatible with no-unused-disable because some of the files are fine as-is.

A better long-term solution would be to apply Prettier to the generated output, similar to what's done in `packages/theme` for Theme artifacts.

* Resolve issues after changing ESLint resolver

Co-authored-by: aduth <aduth@git.wordpress.org>
Co-authored-by: swissspidy <swissspidy@git.wordpress.org>
Co-authored-by: manzoorwanijk <manzoorwanijk@git.wordpress.org>