Add specific classname to block toolbar icon (#73058)
The wrapping button for the block toolbar icon has three different possible classnames and the icon itself had a generic classname by default when using the BlockIcon component. I added a className to the BlockIcon used by the BlockToolbarIcon so it could be targeted in all its instances and be unique from the parent selector. This allows us to target it with CSS more easily and keep styles targeted and tidy.

Co-authored-by: jeryj <jeryj@git.wordpress.org>
Co-authored-by: mikachan <mikachan@git.wordpress.org>