Fix classnames for link preview component  (#73698)
The link control preview appears to have been a search item at one point, or maybe was copied from that. The classnames and purpose have evolved since then, and this is no longer a search item. The classes were wrapped up with the search results, which is a different component. This cleans up and clarifies their separation.