Theme: Generate and expose TypeScript types for token options (#74080)
* Theme: Generate and expose TypeScript types for token options

* UI: Use theme package types instead of hardcoded token types

* Restore explicit padding options to Storybook story

Storybook can't pick it up automatically, likely because of DimensionVariant support.

The numeric values should be removed, as they're no longer supported.

Use `satisfies` to enforce type-checking on valid options.

* Don't auto-sort type union values

This can lead to strange results in the case of t-shirt sizing. While the sort order can't always be guaranteed in how the plugin discovers matches across the full set of tokens, it's a better default behavior.

Co-authored-by: aduth <aduth@git.wordpress.org>
Co-authored-by: mirka <0mirka00@git.wordpress.org>
Co-authored-by: jsnajdr <jsnajdr@git.wordpress.org>